requirements.txt
# pytorch
http://download.pytorch.org/whl/cu80/torch-0.2.0.post3-cp35-cp35m-manylinux1_x86_64.whl 
torchvision==0.1.9
torchtext==0.1.1
visdom

# data
scipy
scikit-learn
numpy
pillow

# utils (others)
colorama
tqdm
lmdb
requests
fake-useragent
